Tests & Exceptions
Everything is running except;
All updates
User; TestRegister & CreatePasswordHash

+++ BestinClass.Core/Application Service/Impl/ReviewService.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using BestinClass.Core.Application_Service.Service;
 using BestinClass.Core.Domain_Service;
@@ -18,6 +19,7 @@ namespace BestinClass.Core.Application_Service.Impl
         public Review NewReview(int carId, string header, string body, int ratingEveryday, int ratingWeekend, int ratingPracticality,
             int ratingExterior, int ratingInterior)
         {
+            
             var review = new Review()
             {
                 CarId = carId,
@@ -36,16 +38,28 @@ namespace BestinClass.Core.Application_Service.Impl
 
         public Review CreateReview(Review review)
         {
+            if(review.Header.Length < 1)
+                { throw new InvalidDataException("To create review, a header must be attached."); }
+            if(review.Body.Length < 1)
+                { throw new InvalidDataException("To create review, a body must be attached."); }
+            if(review.RatingEveryday < 0 || review.RatingExterior < 0 || review.RatingInterior < 0 ||
+                review.RatingOverall < 0 || review.RatingPracticality < 0 || review.RatingWeekend < 0)
+                { throw new InvalidDataException("Review ratings can't be below 0."); }
+
             return _reviewRepository.CreateReview(review);
         }
 
         public List<Review> GetAllReviews()
         {
+            if (_reviewRepository.ReadAllReviews().Count() < 1)
+                { throw new FileNotFoundException("Database is empty."); }
             return _reviewRepository.ReadAllReviews().ToList();
         }
 
         public Review GetReviewById(int id)
         {
+            if(_reviewRepository.GetReviewById(id) == null)
+                { throw new FileNotFoundException("No match were found."); }
             return _reviewRepository.GetReviewById(id);
         }
 
@@ -56,6 +70,8 @@ namespace BestinClass.Core.Application_Service.Impl
 
         public void DeleteReview(int id)
         {
+            if (_reviewRepository.GetReviewById(id) == null)
+                { throw new FileNotFoundException("No match were found."); }
             _reviewRepository.DeleteReview(id);
         }
     }
